requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
black==22.6.0
cfgv==3.3.1
click==8.1.3
coverage==6.4.2
Deprecated==1.2.13
distlib==0.3.4
Django==4.0.6
django-filter==22.1
django-redis==5.2.0
djangorestframework==3.13.1
factory-boy==3.2.1
Faker==13.15.0
filelock==3.7.1
flake8==4.0.1
freezegun==1.2.1
identify==2.5.1
inflection==0.5.1
iniconfig==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.7.0
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.20.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.9
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-factoryboy==2.5.0
pytest-flake8==1.1.1
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
PyYAML==6.0
redis==4.3.4
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing-extensions==4.3.0
virtualenv==20.15.1
wrapt==1.14.1
